+++ cmd/metrics-powerstore/main.go
@@ -10,32 +10,23 @@ package main
 
 import (
 	"context"
-	"errors"
 	"expvar"
 	"fmt"
-	"io/ioutil"
 	"net/http"
-	"path/filepath"
-	"regexp"
 	"runtime"
 	"strconv"
 	"strings"
 	"time"
 
-	csictx "github.com/dell/gocsi/context"
-
 	_ "net/http/pprof"
 
+	"github.com/dell/csm-metrics-powerstore/internal/common"
 	"github.com/dell/csm-metrics-powerstore/internal/entrypoint"
 	"github.com/dell/csm-metrics-powerstore/internal/k8s"
 	"github.com/dell/csm-metrics-powerstore/internal/service"
 	otlexporters "github.com/dell/csm-metrics-powerstore/opentelemetry/exporters"
 	tracer "github.com/dell/csm-metrics-powerstore/opentelemetry/tracers"
-	"github.com/dell/gopowerstore"
 	"github.com/sirupsen/logrus"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
-	"gopkg.in/yaml.v2"
 
 	"os"
 
@@ -52,8 +43,6 @@ const (
 	defaultDebugPort               = "9090"
 	defaultCertFile                = "/certs/localhost.crt"
 	defaultKeyFile                 = "/certs/localhost.key"
-	// EnvThrottlingRateLimit sets a number of concurrent requests to APi
-	EnvThrottlingRateLimit = "X_CSI_POWERSTORE_THROTTLING_RATE_LIMIT"
 )
 
 func main() {
@@ -206,7 +195,7 @@ func updateTracing(logger *logrus.Logger) {
 }
 
 func updatePowerStoreConnection(powerStoreSvc *service.PowerStoreService, logger *logrus.Logger) {
-	arrays, _, _, err := GetPowerStoreArrays(defaultStorageSystemConfigFile, logger)
+	arrays, _, _, err := common.GetPowerStoreArrays(defaultStorageSystemConfigFile, logger)
 	if err != nil {
 		logger.WithError(err).Fatal("initialize arrays in controller service")
 	}
@@ -312,92 +301,4 @@ func updateService(pstoreSvc *service.PowerStoreService, logger *logrus.Logger) 
 	}
 	pstoreSvc.MaxPowerStoreConnections = maxPowerStoreConcurrentRequests
 	logger.WithField("max_connections", maxPowerStoreConcurrentRequests).Debug("setting max powerstore connections")
-}
-
-// GetPowerStoreArrays parses config.yaml file, initializes gopowerstore Clients and composes map of arrays for ease of access.
-// It will return array that can be used as default as a second return parameter.
-// If config does not have any array as a default then the first will be returned as a default.
-func GetPowerStoreArrays(filePath string, logger *logrus.Logger) (map[string]*service.PowerStoreArray, map[string]string, *service.PowerStoreArray, error) {
-	type config struct {
-		Arrays []*service.PowerStoreArray `yaml:"arrays"`
-	}
-
-	data, err := ioutil.ReadFile(filepath.Clean(filePath))
-	if err != nil {
-		logger.WithError(err).Errorf("cannot read file %s", filePath)
-		return nil, nil, nil, err
-	}
-
-	var cfg config
-	err = yaml.Unmarshal(data, &cfg)
-	if err != nil {
-		logger.WithError(err).Errorf("cannot unmarshal data")
-		return nil, nil, nil, err
-	}
-
-	arrayMap := make(map[string]*service.PowerStoreArray)
-	mapper := make(map[string]string)
-	var defaultArray *service.PowerStoreArray
-	foundDefault := false
-
-	if len(cfg.Arrays) == 0 {
-		return arrayMap, mapper, defaultArray, nil
-	}
-
-	// Safeguard if user doesn't set any array as default, we just use first one
-	defaultArray = cfg.Arrays[0]
-
-	// Convert to map for convenience and init gopowerstore.Client
-	for _, array := range cfg.Arrays {
-		array := array
-		if array == nil {
-			return arrayMap, mapper, defaultArray, nil
-		}
-		if array.GlobalId == "" {
-			return nil, nil, nil, errors.New("No GlobalID field found in config.yaml. Update config.yaml according to the documentation.")
-		}
-		clientOptions := gopowerstore.NewClientOptions()
-		clientOptions.SetInsecure(array.Insecure)
-
-		if throttlingRateLimit, ok := csictx.LookupEnv(context.Background(), EnvThrottlingRateLimit); ok {
-			rateLimit, err := strconv.Atoi(throttlingRateLimit)
-			if err != nil {
-				logger.Errorf("can't get throttling rate limit, using default")
-			} else {
-				clientOptions.SetRateLimit(uint64(rateLimit))
-			}
-		}
-
-		c, err := gopowerstore.NewClientWithArgs(
-			array.Endpoint, array.Username, array.Password, clientOptions)
-		if err != nil {
-			return nil, nil, nil, status.Errorf(codes.FailedPrecondition,
-				"unable to create PowerStore client: %s", err.Error())
-		}
-		array.Client = c
-
-		ips := GetIPListFromString(array.Endpoint)
-		if ips == nil {
-			return nil, nil, nil, fmt.Errorf("can't get ips from endpoint: %s", array.Endpoint)
-		}
-
-		ip := ips[0]
-		array.IP = ip
-		logger.Infof("%s,%s,%s,%s,%t,%t,%s", array.Endpoint, array.GlobalId, array.Username, array.NasName, array.Insecure, array.IsDefault, array.BlockProtocol)
-		arrayMap[array.GlobalId] = array
-		mapper[ip] = array.GlobalId
-		if array.IsDefault && !foundDefault {
-			defaultArray = array
-			foundDefault = true
-		}
-	}
-
-	return arrayMap, mapper, defaultArray, nil
-}
-
-// GetIPListFromString returns list of ips in string form found in input string
-// A return value of nil indicates no match
-func GetIPListFromString(input string) []string {
-	re := regexp.MustCompile(`(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)(\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)){3}`)
-	return re.FindAllString(input, -1)
 }
