+++ internal/common/testdata/invalid-format.yaml
@@ -1,0 +1,1 @@
+invalid_format
