+++ internal/common/testdata/no-global-id.yaml
@@ -1,0 +1,8 @@
+arrays:
+  - endpoint: "https://10.0.0.1/api/rest"
+    username: "user"
+    password: "password"
+    skipCertificateValidation: true
+    isDefault: true
+    blockProtocol: "auto"
+    nasName: "nas-server"
